chore: move vectors to root in assets folder
