+++ web/src/pages/NotFound/index.tsx
@@ -1,4 +1,4 @@
-import Icon404 from "../../assets/vectors/404.svg"
+import Icon404 from "../../assets/404.svg"
 
 export default function NotFound() {
     return (
